Update FindPage location
