+++ shared/common/src/pages/auth/find/Pagination/ChangePasswordSuccess/style.ts
