+++ shared/common/src/pages/auth/find/Pagination/EmailCheck/style.ts
