+++ shared/common/src/pages/auth/find/Pagination/EmailValid/style.ts
