+++ shared/common/src/pages/auth/find/Pagination/NewPassword/index.tsx
