+++ shared/common/src/pages/auth/find/Pagination/NewPassword/style.ts
