+++ shared/common/src/pages/auth/find/style.ts
