+++ shared/common/src/pages/auth/index.ts
@@ -1,2 +1,3 @@
 export { default as LoginPage } from './login'
 export { default as SignUpPage } from './signUp'
+export { default as FindPage } from './find'
