+++ shared/common/src/pages/index.ts
@@ -1,9 +1,8 @@
 export { default as ActivityDetailPage } from './activity/detail'
 export { default as ActivityListPage } from './activity/list'
-export { LoginPage, SignUpPage } from './auth'
+export { LoginPage, SignUpPage, FindPage } from './auth'
 export { default as ClubDetailPage } from './club/detail'
 export { default as StudentPage } from './club/student'
-export { default as FindPage } from './find'
 export { default as HomePage } from './home'
 export { InquiryDetailPage, InquiryPage } from './inquiry'
 export {
